+++ x/auth/ante/expected_keepers.go
@@ -13,6 +13,7 @@ type AccountKeeper interface {
 	SetAccount(ctx sdk.Context, acc types.AccountI)
 	GetModuleAddress(moduleName string) sdk.AccAddress
 	NewAccountWithAddress(ctx sdk.Context, addr sdk.AccAddress) types.AccountI
+	HasAccount(ctx sdk.Context, addr sdk.AccAddress) bool
 }
 
 // FeegrantKeeper defines the expected feegrant keeper.