+++ x/zkauth/types/zkauth_verify_test.go
@@ -64,7 +64,7 @@ func TestVerifyZKAuthSignature(t *testing.T) {
 		ZkAuthSignature: ZKAuthSignature{
 			ZkAuthInputs: &ZKAuthInputs{
 				ProofPoints:  []byte(proofStr),
-				IssF:         "aHR0cHM6Ly9hY2NvdW50cy5nb29nbGUuY29t",
+				IssBase64:    "aHR0cHM6Ly9hY2NvdW50cy5nb29nbGUuY29t",
 				HeaderBase64: "eyJhbGciOiJSUzI1NiIsImtpZCI6IjU1YzE4OGE4MzU0NmZjMTg4ZTUxNTc2YmE3MjgzNmUwNjAwZThiNzMiLCJ0eXAiOiJKV1QifQ",
 				AddressSeed:  "15035161560159971633800983619931498696152633426768016966057770643262022096073",
 			},